# swagger.yaml
openapi: 3.0.1
info:
  title: Office Manager's API
  version: v1
paths:
  /v1/cases:search:
    get:
      tags:
        - Cases
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/CaseRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CaseRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CaseResponse'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CaseResponse'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CaseResponse'
  /v1/cases/{caseId}:
    get:
      tags:
        - Cases
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/CaseResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/CaseResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/CaseResponse'
  /v1/cases:
    get:
      tags:
        - Cases
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CaseResponse'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CaseResponse'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CaseResponse'
    post:
      tags:
        - Cases
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/CaseRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CaseRequest'
      responses:
        '200':
          description: OK
    put:
      tags:
        - Cases
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/CaseRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CaseRequest'
      responses:
        '200':
          description: OK
    delete:
      tags:
        - Cases
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
  /v1/client-meetings/{clientMeetingId}:
    get:
      tags:
        - ClientMeetings
      parameters:
        - name: clientMeetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ClientMeetingResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ClientMeetingResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ClientMeetingResponse'
  /v1/client-meetings:
    get:
      tags:
        - ClientMeetings
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientMeetingResponse'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientMeetingResponse'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientMeetingResponse'
    post:
      tags:
        - ClientMeetings
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientMeetingRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/ClientMeetingRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ClientMeetingRequest'
      responses:
        '200':
          description: OK
    put:
      tags:
        - ClientMeetings
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientMeetingRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/ClientMeetingRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ClientMeetingRequest'
      responses:
        '200':
          description: OK
    delete:
      tags:
        - ClientMeetings
      parameters:
        - name: clientMeetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
  /v1/client-meetings/client/{clientId}:
    get:
      tags:
        - ClientMeetings
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientMeetingResponse'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientMeetingResponse'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientMeetingResponse'
  /v1/clients/office/{officeId}:
    get:
      tags:
        - Clients
      parameters:
        - name: officeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientResponse'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientResponse'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientResponse'
  /v1/clients:paginated:
    get:
      tags:
        - Clients
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientPaginationRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/ClientPaginationRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ClientPaginationRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ClientResponsePaginationResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponsePaginationResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ClientResponsePaginationResponse'
  /v1/clients/{clientId}:
    get:
      tags:
        - Clients
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ClientResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
  /v1/clients:
    get:
      tags:
        - Clients
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientResponse'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientResponse'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientResponse'
    post:
      tags:
        - Clients
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/ClientRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ClientRequest'
      responses:
        '200':
          description: OK
    put:
      tags:
        - Clients
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/ClientRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ClientRequest'
      responses:
        '200':
          description: OK
    delete:
      tags:
        - Clients
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
components:
  schemas:
    CaseRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        clientId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/CaseStatus'
      additionalProperties: false
    CaseResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        client:
          $ref: '#/components/schemas/ClientResponse'
        status:
          $ref: '#/components/schemas/CaseStatus'
        creationDate:
          type: string
          format: date-time
        conclusionDate:
          type: string
          format: date-time
      additionalProperties: false
    CaseStatus:
      enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 7
        - 8
        - 9
      type: integer
      format: int32
    ClientMeetingRequest:
      type: object
      properties:
        meetingDate:
          type: string
          format: date-time
        description:
          type: string
          nullable: true
        clientId:
          type: string
          format: uuid
      additionalProperties: false
    ClientMeetingResponse:
      type: object
      additionalProperties: false
    ClientPaginationRequest:
      type: object
      properties:
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        officeId:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        cpf:
          type: string
          nullable: true
      additionalProperties: false
    ClientRequest:
      type: object
      properties:
        cpf:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        officeId:
          type: string
          format: uuid
      additionalProperties: false
    ClientResponse:
      type: object
      properties:
        cpf:
          type: string
          nullable: true
          readOnly: true
        name:
          type: string
          nullable: true
          readOnly: true
        address:
          type: string
          nullable: true
          readOnly: true
        phone:
          type: string
          nullable: true
          readOnly: true
        email:
          type: string
          nullable: true
          readOnly: true
        meetings:
          type: array
          items:
            $ref: '#/components/schemas/ClientMeetingResponse'
          nullable: true
          readOnly: true
      additionalProperties: false
    ClientResponsePaginationResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ClientResponse'
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/ClientResponse'
          nullable: true
        currentPage:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
        totalItems:
          type: integer
          format: int32
      additionalProperties: false
...